% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_formatted_string.R
\name{create_formatted_string}
\alias{create_formatted_string}
\title{Title}
\usage{
create_formatted_string(fastq_data, display_sequence_id = TRUE,
  display_sequence = TRUE, display_emoji = TRUE,
  display_encoded_quality = FALSE, display_decoded_quality = FALSE)
}
\arguments{
\item{display_decoded_quality}{}
}
\description{
Title
}
