% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore-basecaller-and-fast5type.R
\name{explore_basecaller_and_fast5type}
\alias{explore_basecaller_and_fast5type}
\title{Explore a fast5 file to find parameters of the experiment}
\usage{
explore_basecaller_and_fast5type(fast5file_path, basecall_group)
}
\arguments{
\item{fast5file_path}{a character string. Path of a fast5file for determining
parameter of the experiment.}

\item{basecall_group}{a character string. Name of the level
in the Fast5 file hierarchy from which to read the data e.g. "Basecall_1D_000"}
}
\value{
A list containing the all the relevant data
}
\description{
Explore a fast5 file to find parameters of the experiment
}
