% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_space.R
\name{insert_space}
\alias{insert_space}
\title{bs
bs2}
\usage{
insert_space(character_string, num_spaces)
}
\arguments{
\item{character_string}{cs}

\item{num_spaces}{ns}
}
\value{
a val
}
\description{
bs
bs2
}
