% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-plot.R
\name{make_plot}
\alias{make_plot}
\title{Title}
\usage{
make_plot(read_data, width, height, z_normalize = FALSE,
  truncate_spikes = FALSE, highlight = FALSE,
  highlight_region = NULL)
}
\arguments{
\item{truncate_spikes}{}
}
\description{
Title
}
