% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/truncate-spikes.R
\name{truncate_data_spikes}
\alias{truncate_data_spikes}
\title{Truncate spikes}
\usage{
truncate_data_spikes(data, data_mean, data_sd, threshold = 2)
}
\arguments{
\item{data}{Data that needs its spikes to be truncated}

\item{data_mean}{dm}

\item{data_sd}{dsd}

\item{threshold}{t}
}
\value{
Same data as input but with spikes clipped
}
\description{
It has been observed that the signal contains a lot of high-voltage spikes
distributed throughout the lenght of the read. This function truncates/clips
these spikes to a specified threshold value
}
